requirements.txt
discord.py>=1.5
python-dotenv